CI-629 altered ukprn to long

+++ src/Sfa.Das.ApprenticeshipInfoService.Application/Models/OrganisationDocument.cs
@@ -20,6 +20,6 @@
 
         public string EpaOrganisationIdentifierKeyword { get; set; }
 
-        public int? Ukprn { get; set; }
+        public long? Ukprn { get; set; }
     }
 }
